made a working clock and had some more routing done for the assignment
